Setup of entity segments
- Added entity segment interface
- Added entity tags
- Created multiple entity segment implementations
