+++ Common/src/main/java/dk/sdu/mmmi/cbse/common/data/EntityTag.java
@@ -1,0 +1,9 @@
+package dk.sdu.mmmi.cbse.common.data;
+
+public enum EntityTag {
+    UNTAGGED,
+    PLAYER,
+    ENEMY,
+    ASTEROID,
+    BULLET
+}
