+++ Common/src/main/java/dk/sdu/mmmi/cbse/common/entitysegments/CircleColliderSegment.java
@@ -1,0 +1,19 @@
+package dk.sdu.mmmi.cbse.common.entitysegments;
+
+import dk.sdu.mmmi.cbse.common.data.Entity;
+import dk.sdu.mmmi.cbse.common.data.GameData;
+
+public class CircleColliderSegment implements EntitySegment{
+    private double radius;
+    public void setRadius(double radius) {
+        this.radius = radius;
+    }
+    public double getRadius() {
+        return radius;
+    }
+
+    @Override
+    public void process(GameData gameData, Entity entity) {
+
+    }
+}
