+++ Common/src/main/java/dk/sdu/mmmi/cbse/common/entitysegments/EntitySegment.java
@@ -1,0 +1,8 @@
+package dk.sdu.mmmi.cbse.common.entitysegments;
+
+import dk.sdu.mmmi.cbse.common.data.Entity;
+import dk.sdu.mmmi.cbse.common.data.GameData;
+
+public interface EntitySegment {
+    void process(GameData gameData, Entity entity);
+}
