+++ Common/src/main/java/dk/sdu/mmmi/cbse/common/entitysegments/FiringSegment.java
@@ -1,0 +1,11 @@
+package dk.sdu.mmmi.cbse.common.entitysegments;
+
+import dk.sdu.mmmi.cbse.common.data.Entity;
+import dk.sdu.mmmi.cbse.common.data.GameData;
+
+public class FiringSegment implements EntitySegment{
+    @Override
+    public void process(GameData gameData, Entity entity) {
+
+    }
+}
