+++ Common/src/main/java/module-info.java
@@ -1,4 +1,5 @@
 module Common {
     exports dk.sdu.mmmi.cbse.common.services;
     exports dk.sdu.mmmi.cbse.common.data;
+    exports dk.sdu.mmmi.cbse.common.entitysegments;
 }
