--- CommonPlayer/src/main/java/module-info.java
@@ -1,4 +1,0 @@
-module CommonPlayer {
-    requires Common;
-    exports dk.sdu.mmmi.cbse.common.player;
-}
