+++ Enemy/src/main/java/dk/sdu/mmmi/cbse/enemysystem/EnemyPlugin.java
@@ -4,9 +4,7 @@ import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
 import dk.sdu.mmmi.cbse.common.data.Vector;
 import dk.sdu.mmmi.cbse.common.data.World;
-import dk.sdu.mmmi.cbse.common.player.Player;
 import dk.sdu.mmmi.cbse.common.services.IGamePluginService;
-import dk.sdu.mmmi.cbse.enemysystem.Enemy;
 
 public class EnemyPlugin implements IGamePluginService {
