+++ pom.xml
@@ -19,7 +19,6 @@
         <module>Enemy</module>
         <module>Asteroid</module>
         <module>Collision</module>
-        <module>CommonPlayer</module>
     </modules>
     <properties>
         <java>java</java>
